requirements.txt
Pillow==8.1.0
python-docx==0.8.10
requests==2.25.1